requirements.txt
absl-py==2.4.0
altgraph==0.17.5
attrs==25.4.0
cffi==2.0.0
comtypes==1.4.15
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.9
MouseInfo==0.1.3
numpy==2.4.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
protobuf==3.20.3
psutil==7.2.2
PyAutoGUI==0.9.54
pycaw==20220416
pycparser==3.0
PyGetWindow==0.0.9
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pystray==0.19.5
python-dateutil==2.9.0.post0
pytweening==1.2.0
pywin32-ctypes==0.2.3
six==1.17.0
sounddevice==0.5.5